Organize core and feature packages and add CI
