+++ .github/workflows/android.yml
@@ -1,0 +1,23 @@
+name: Android CI
+
+on:
+  push:
+    branches: [ master, feature/dashboard-concepts ]
+  pull_request:
+    branches: [ master ]
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - uses: actions/setup-java@v4
+        with:
+          distribution: 'temurin'
+          java-version: '17'
+      - name: Grant execute permission for gradlew
+        run: chmod +x gradlew
+      - name: Run lint
+        run: ./gradlew lint
+      - name: Assemble Debug
+        run: ./gradlew assembleDebug
