+++ app/src/main/java/com/concepts_and_quizzes/cds/MainActivity.kt
@@ -10,21 +10,6 @@ import androidx.compose.animation.fadeOut
 import androidx.compose.animation.slideInHorizontally
 import androidx.compose.animation.slideOutHorizontally
 import androidx.compose.animation.togetherWith
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Button
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.unit.dp
 import androidx.core.content.edit
 import androidx.security.crypto.EncryptedSharedPreferences
 import androidx.security.crypto.MasterKey
@@ -32,6 +17,8 @@ import com.concepts_and_quizzes.cds.auth.AuthViewModel
 import com.concepts_and_quizzes.cds.auth.AuthViewModelFactory
 import com.concepts_and_quizzes.cds.auth.LoginScreen
 import com.concepts_and_quizzes.cds.auth.RegisterScreen
+import com.concepts_and_quizzes.cds.core.theme.CDSTheme
+import com.concepts_and_quizzes.cds.ui.dashboard.DashboardScreen
 
 class MainActivity : ComponentActivity() {
     private val viewModel: AuthViewModel by viewModels { AuthViewModelFactory(this) }
@@ -57,45 +44,47 @@ class MainActivity : ComponentActivity() {
         if (viewModel.currentUser == null) viewModel.trySilentSignIn()
 
         setContent {
-            val currentUser = viewModel.currentUser
-            val showRegister = viewModel.showRegister
-            val needsAuth = currentUser == null && !skipAuthForTesting
+            CDSTheme {
+                val currentUser = viewModel.currentUser
+                val showRegister = viewModel.showRegister
+                val needsAuth = currentUser == null && !skipAuthForTesting
 
-            AnimatedContent(targetState = needsAuth, label = "auth") { needsAuth ->
-                if (needsAuth) {
-                    AnimatedContent(
-                        targetState = showRegister,
-                        transitionSpec = {
-                            slideInHorizontally { it } + fadeIn() togetherWith
-                                slideOutHorizontally { -it } + fadeOut()
-                        },
-                        label = "form"
-                    ) { reg ->
-                        if (reg) {
-                            RegisterScreen(
-                                viewModel = viewModel,
-                                onRegistrationSuccess = { email, password ->
-                                    cacheEmailPassword(email, password)
-                                    viewModel.toggleForm()
-                                },
-                                onBackToLogin = { viewModel.toggleForm() }
-                            )
-                        } else {
-                            LoginScreen(
-                                viewModel = viewModel,
-                                onNavigateToRegister = { viewModel.toggleForm() },
-                                onLoginSuccess = { email, password -> cacheEmailPassword(email, password) }
-                            )
+                AnimatedContent(targetState = needsAuth, label = "auth") { needsAuth ->
+                    if (needsAuth) {
+                        AnimatedContent(
+                            targetState = showRegister,
+                            transitionSpec = {
+                                slideInHorizontally { it } + fadeIn() togetherWith
+                                    slideOutHorizontally { -it } + fadeOut()
+                            },
+                            label = "form"
+                        ) { reg ->
+                            if (reg) {
+                                RegisterScreen(
+                                    viewModel = viewModel,
+                                    onRegistrationSuccess = { email, password ->
+                                        cacheEmailPassword(email, password)
+                                        viewModel.toggleForm()
+                                    },
+                                    onBackToLogin = { viewModel.toggleForm() }
+                                )
+                            } else {
+                                LoginScreen(
+                                    viewModel = viewModel,
+                                    onNavigateToRegister = { viewModel.toggleForm() },
+                                    onLoginSuccess = { email, password -> cacheEmailPassword(email, password) }
+                                )
+                            }
                         }
+                    } else {
+                        DashboardScreen(
+                            name = currentUser?.displayName ?: "",
+                            onSignOut = {
+                                viewModel.signOut()
+                                prefs.edit { clear() }
+                            }
+                        )
                     }
-                } else {
-                    DashboardScreen(
-                        name = currentUser?.displayName ?: "",
-                        onSignOut = {
-                            viewModel.signOut()
-                            prefs.edit { clear() }
-                        }
-                    )
                 }
             }
         }
@@ -105,29 +94,6 @@ class MainActivity : ComponentActivity() {
         prefs.edit {
             putString("email", email)
             putString("password", password)
-        }
-    }
-}
-
-@Composable
-fun DashboardScreen(name: String, onSignOut: () -> Unit) {
-    Scaffold { padding ->
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(padding)
-                .padding(32.dp),
-            verticalArrangement = Arrangement.Center,
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            Text(
-                text = stringResource(id = R.string.welcome_message, name),
-                style = MaterialTheme.typography.headlineMedium,
-            )
-            Spacer(modifier = Modifier.height(24.dp))
-            Button(onClick = onSignOut) {
-                Text(text = stringResource(id = R.string.sign_out))
-            }
         }
     }
 }
