+++ app/src/main/java/com/concepts_and_quizzes/cds/core/components/AppBar.kt
@@ -1,0 +1,10 @@
+package com.concepts_and_quizzes.cds.core.components
+
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
+import androidx.compose.runtime.Composable
+
+@Composable
+fun CdsAppBar(title: String) {
+    TopAppBar(title = { Text(text = title) })
+}
