+++ app/src/main/java/com/concepts_and_quizzes/cds/core/components/BottomNavBar.kt
@@ -1,0 +1,9 @@
+package com.concepts_and_quizzes.cds.core.components
+
+import androidx.compose.material3.NavigationBar
+import androidx.compose.runtime.Composable
+
+@Composable
+fun CdsBottomNavBar() {
+    NavigationBar { }
+}
