+++ app/src/main/java/com/concepts_and_quizzes/cds/ui/concepts/ConceptsScreen.kt
@@ -1,0 +1,9 @@
+package com.concepts_and_quizzes.cds.ui.concepts
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun ConceptsScreen() {
+    Text(text = "Concepts")
+}
