+++ app/src/main/java/com/concepts_and_quizzes/cds/ui/english/EnglishScreen.kt
@@ -1,0 +1,9 @@
+package com.concepts_and_quizzes.cds.ui.english
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun EnglishScreen() {
+    Text(text = "English")
+}
